Change `Tsn` from `u32` to `u64` and update serialization, deserialization, size calculations, logic, and tests across `mvcc_node_free_list`, `mvcc_node_header`, and related modules. Update assertions for modified page counts in `mvcc.rs`.

+++ src/mvcc_common.rs
@@ -54,4 +54,4 @@ pub type Result<T> = std::result::Result<T, LmdbError>;
 pub struct PageID(pub u32);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Tsn(pub u32);
+pub struct Tsn(pub u64);
